实装：（能源）实装：（能源）实装了能源管理系统的人力发电功能 调整：优化了部分代码中的变量数据格式

+++ Script/Design/attr_calculation.py
@@ -662,7 +662,7 @@ def get_juel(value: int) -> int:
     return juel
 
 
-def get_pain_adjust(value: int, level_flag = False) -> int:
+def get_pain_adjust(value: int, level_flag = False) -> float:
     """
     按润滑程度修正苦痛值比例
     Keyword arguments:
